--- js/functions.js
@@ -1,30 +1,0 @@
-const isStringOverMaxLength = (string, maxLength) => string.length <= maxLength;
-
-isStringOverMaxLength('проверяемая строка', 20);
-
-const isPalindrome = (string) => {
-  const normalizedString = string.replaceAll(' ', '').toUpperCase();
-  let reversedString = '';
-  for (let i = normalizedString.length - 1; i >= 0; i--) {
-    reversedString += normalizedString[i];
-  }
-  return normalizedString === reversedString;
-};
-
-isPalindrome('Кекс');
-
-const extractNumbers = (string) => {
-  string = string.toString();
-  let result = '';
-  let number;
-  for (let i = 0; i < string.length; i++) {
-    number = parseInt(string[i], 10);
-    if (!Number.isNaN(number)) {
-      result += number;
-    }
-  }
-
-  return Number(result) || NaN;
-};
-
-extractNumbers('2023 год');
